requirements.txt
flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
flask-cors==4.0.0
pytz==2024.1 
psutil==5.9.8
